requirements.txt
Django==1.11.13
djangorestframework==3.6.3
whitenoise-3.3.1